Move About component to pages folder.
